+++ snug/pagination.py
@@ -6,22 +6,11 @@ import abc
 import typing as t
 from operator import attrgetter
 
-from .compat import PY3
 from .query import Query, async_executor, executor
 
-__all__ = [
-    'paginated',
-    'Page',
-    'Pagelike',
-]
+__all__ = ["paginated", "Page", "Pagelike"]
 
-T = t.TypeVar('T')
-
-if PY3:
-    from ._async import AsyncPaginator
-    _AsyncIterator = t.AsyncIterator
-else:  # pragma: no cover
-    _AsyncIterator = t.Iterator
+T = t.TypeVar("T")
 
 
 class Pagelike(t.Generic[T]):
@@ -40,6 +29,7 @@ class Pagelike(t.Generic[T]):
     For example: ``Pagelike[List[str]]`` indicates a page-like
     object whose ``content`` is a list of strings.
     """
+
     __slots__ = ()
 
     @abc.abstractproperty
@@ -76,19 +66,20 @@ class Page(Pagelike[T]):
     next_query: ~snug.Query[Pagelike[T]]] or None
         The query to retrieve the next page.
     """
-    __slots__ = '_content', '_next_query'
+
+    __slots__ = "_content", "_next_query"
 
     def __init__(self, content, next_query=None):
         self._content, self._next_query = content, next_query
 
-    content = property(attrgetter('_content'))
-    next_query = property(attrgetter('_next_query'))
+    content = property(attrgetter("_content"))
+    next_query = property(attrgetter("_next_query"))
 
     def __repr__(self):
-        return 'Page({})'.format(self._content)
+        return "Page({})".format(self._content)
 
 
-class paginated(Query[t.Union[t.Iterator[T], _AsyncIterator[T]]]):
+class paginated(Query[t.Union[t.Iterator[T], t.AsyncIterator[T]]]):
     """A paginated version of a query.
     Executing it returns an :term:`iterator`
     or :term:`async iterator <asynchronous iterator>`.
@@ -126,7 +117,8 @@ class paginated(Query[t.Union[t.Iterator[T], _AsyncIterator[T]]]):
         async for foo in execute_async(query):  # python 3.5.2+ only
             ...
     """
-    __slots__ = '_query'
+
+    __slots__ = "_query"
 
     def __init__(self, query):
         self._query = query
@@ -141,35 +133,31 @@ class paginated(Query[t.Union[t.Iterator[T], _AsyncIterator[T]]]):
         """
         return Paginator(self._query, executor(client=client, auth=auth))
 
-    if PY3:
-        def __execute_async__(self, client, auth):
-            """Execute the paginated query asynchronously.
+    def __execute_async__(self, client, auth):
+        """Execute the paginated query asynchronously.
 
-            Note
-            ----
-            This method does not need to be awaited.
+        Note
+        ----
+        This method does not need to be awaited.
 
-            Returns
-            -------
-            ~typing.AsyncIterator[T]
-                An asynchronous iterator yielding page content.
-            """
-            return AsyncPaginator(self._query,
-                                  async_executor(client=client, auth=auth))
-
-    else:  # pragma: no cover
-        def __execute_async__(self, client, auth):
-            raise NotImplementedError(
-                'async execution of paginated queries is python 3.5.2+ only')
+        Returns
+        -------
+        ~typing.AsyncIterator[T]
+            An asynchronous iterator yielding page content.
+        """
+        return AsyncPaginator(
+            self._query, async_executor(client=client, auth=auth)
+        )
 
     def __repr__(self):
-        return 'paginated({})'.format(self._query)
+        return "paginated({})".format(self._query)
 
 
 class Paginator(t.Iterator[T]):
     """An iterator which keeps executing the next query in the page sequece,
     returning the page content."""
-    __slots__ = '_executor', '_next_query'
+
+    __slots__ = "_executor", "_next_query"
 
     def __init__(self, next_query, executor):
         self._next_query, self._executor = next_query, executor
@@ -184,5 +172,23 @@ class Paginator(t.Iterator[T]):
         self._next_query = page.next_query
         return page.content
 
-    if not PY3:  # pragma: no cover
-        next = __next__
+
+class AsyncPaginator(t.AsyncIterator[T]):
+    """An async iterator which keeps executing
+    the next query in the page sequence"""
+
+    __slots__ = "_executor", "_next_query"
+
+    def __init__(self, next_query, executor):
+        self._next_query, self._executor = next_query, executor
+
+    def __aiter__(self):
+        return self
+
+    async def __anext__(self):
+        """the content of the next page"""
+        if self._next_query is None:
+            raise StopAsyncIteration()
+        page = await self._executor(self._next_query)
+        self._next_query = page.next_query
+        return page.content