+++ tests/conftest.py
@@ -1,14 +1,19 @@
+import asyncio
+
 import pytest
 
 
 def pytest_addoption(parser):
-    parser.addoption("--live", action="store_true",
-                     default=False, help="run against live data")
+    parser.addoption(
+        "--live",
+        action="store_true",
+        default=False,
+        help="run against live data",
+    )
 
 
 @pytest.fixture
-def loop(scope='session'):
-    asyncio = pytest.importorskip('asyncio')
+def loop(scope="session"):
     return asyncio.get_event_loop()