+++ tests/test_query.py
@@ -1,19 +1,28 @@
+import asyncio
 import inspect
-import sys
+import urllib.request
 from operator import methodcaller
-
-import pytest
-from gentools import py2_compatible, return_
 
 import snug
 
-try:
-    import urllib.request as urllib
-except ImportError:
-    import urllib2 as urllib
+
+async def awaitable(obj):
+    """an awaitable returning given object"""
+    await asyncio.sleep(0)
+    return obj
 
 
-py3 = pytest.mark.skipif(sys.version_info < (3, ), reason='python 3+ only')
+class MockAsyncClient:
+    def __init__(self, response):
+        self.response = response
+
+    async def send(self, req):
+        await asyncio.sleep(0)
+        self.request = req
+        return self.response
+
+
+snug.send_async.register(MockAsyncClient, MockAsyncClient.send)
 
 
 class MockClient(object):
@@ -29,7 +38,6 @@ snug.send.register(MockClient, MockClient.send)
 
 
 def test__execute__():
-
     class StringClient:
         def __init__(self, mappings):
             self.mappings = mappings
@@ -39,189 +47,178 @@ def test__execute__():
 
     snug.send.register(StringClient, StringClient.send)
 
-    client = StringClient({
-        'foo/posts/latest': 'redirect:/posts/latest/',
-        'foo/posts/latest/': 'redirect:/posts/december/',
-        'foo/posts/december/': b'hello world'
-    })
+    client = StringClient(
+        {
+            "foo/posts/latest": "redirect:/posts/latest/",
+            "foo/posts/latest/": "redirect:/posts/december/",
+            "foo/posts/december/": b"hello world",
+        }
+    )
 
     class MyQuery(object):
-        @py2_compatible
         def __iter__(self):
-            redirect = yield '/posts/latest'
-            redirect = yield redirect.split(':')[1]
-            response = yield redirect.split(':')[1]
-            return_(response.decode('ascii'))
+            redirect = yield "/posts/latest"
+            redirect = yield redirect.split(":")[1]
+            response = yield redirect.split(":")[1]
+            return response.decode("ascii")
 
-    assert snug.query._default_execute_method(
-        MyQuery(),
-        client,
-        lambda s: 'foo' + s) == 'hello world'
+    assert (
+        snug.Query.__execute__(MyQuery(), client, lambda s: "foo" + s)
+        == "hello world"
+    )
 
 
-@py2_compatible
 def myquery():
-    return_((yield snug.GET('my/url')))
+    return (yield snug.GET("my/url"))
 
 
 class TestExecute:
-
     def test_defaults(self, mocker):
-        send = mocker.patch('snug.query.send', autospec=True)
+        send = mocker.patch("snug.query.send", autospec=True)
 
         assert snug.execute(myquery()) == send.return_value
         client, req = send.call_args[0]
-        assert isinstance(client, urllib.OpenerDirector)
-        assert req == snug.GET('my/url')
+        assert isinstance(client, urllib.request.OpenerDirector)
+        assert req == snug.GET("my/url")
 
     def test_custom_client(self):
         client = MockClient(snug.Response(204))
 
         result = snug.execute(myquery(), client=client)
         assert result == snug.Response(204)
-        assert client.request == snug.GET('my/url')
+        assert client.request == snug.GET("my/url")
 
     def test_custom_execute(self):
         client = MockClient(snug.Response(204))
 
         class MyQuery(object):
             def __execute__(self, client, auth):
-                return client.send(snug.GET('my/url'))
+                return client.send(snug.GET("my/url"))
 
         result = snug.execute(MyQuery(), client=client)
         assert result == snug.Response(204)
-        assert client.request == snug.GET('my/url')
+        assert client.request == snug.GET("my/url")
 
     def test_auth(self):
         client = MockClient(snug.Response(204))
 
-        result = snug.execute(myquery(),
-                              auth=('user', 'pw'),
-                              client=client)
+        result = snug.execute(myquery(), auth=("user", "pw"), client=client)
         assert result == snug.Response(204)
         assert client.request == snug.GET(
-            'my/url', headers={'Authorization': 'Basic dXNlcjpwdw=='})
+            "my/url", headers={"Authorization": "Basic dXNlcjpwdw=="}
+        )
 
     def test_none_auth(self):
         client = MockClient(snug.Response(204))
 
-        result = snug.execute(myquery(),
-                              auth=None,
-                              client=client)
+        result = snug.execute(myquery(), auth=None, client=client)
         assert result == snug.Response(204)
-        assert client.request == snug.GET('my/url')
+        assert client.request == snug.GET("my/url")
 
     def test_auth_callable(self):
         client = MockClient(snug.Response(204))
-        auther = methodcaller('with_headers', {'X-My-Auth': 'letmein'})
+        auther = methodcaller("with_headers", {"X-My-Auth": "letmein"})
 
-        result = snug.execute(myquery(),
-                              auth=auther,
-                              client=client)
+        result = snug.execute(myquery(), auth=auther, client=client)
         assert result == snug.Response(204)
         assert client.request == snug.GET(
-            'my/url', headers={'X-My-Auth': 'letmein'})
+            "my/url", headers={"X-My-Auth": "letmein"}
+        )
 
 
-@py3
 class TestExecuteAsync:
-
     def test_defaults(self, loop, mocker):
-        import asyncio
-        from .py3_only import awaitable
-
-        send = mocker.patch('snug._async.send_async',
-                            return_value=awaitable(snug.Response(204)))
+        send = mocker.patch(
+            "snug.query.send_async", return_value=awaitable(snug.Response(204))
+        )
 
         future = snug.execute_async(myquery())
         result = loop.run_until_complete(future)
         assert result == snug.Response(204)
         client, req = send.call_args[0]
         assert isinstance(client, asyncio.AbstractEventLoop)
-        assert req == snug.GET('my/url')
+        assert req == snug.GET("my/url")
 
     def test_custom_client(self, loop):
-        from .py3_only import MockAsyncClient
+
         client = MockAsyncClient(snug.Response(204))
 
         future = snug.execute_async(myquery(), client=client)
         result = loop.run_until_complete(future)
         assert result == snug.Response(204)
-        assert client.request == snug.GET('my/url')
+        assert client.request == snug.GET("my/url")
 
     def test_custom_execute(self, loop):
-        from .py3_only import MockAsyncClient
+
         client = MockAsyncClient(snug.Response(204))
 
         class MyQuery:
             def __execute_async__(self, client, auth):
-                return client.send(snug.GET('my/url'))
+                return client.send(snug.GET("my/url"))
 
         future = snug.execute_async(MyQuery(), client=client)
         result = loop.run_until_complete(future)
         assert result == snug.Response(204)
-        assert client.request == snug.GET('my/url')
+        assert client.request == snug.GET("my/url")
 
     def test_auth(self, loop):
-        from .py3_only import MockAsyncClient
+
         client = MockAsyncClient(snug.Response(204))
 
-        future = snug.execute_async(myquery(),
-                                    auth=('user', 'pw'),
-                                    client=client)
+        future = snug.execute_async(
+            myquery(), auth=("user", "pw"), client=client
+        )
         result = loop.run_until_complete(future)
         assert result == snug.Response(204)
         assert client.request == snug.GET(
-            'my/url', headers={'Authorization': 'Basic dXNlcjpwdw=='})
+            "my/url", headers={"Authorization": "Basic dXNlcjpwdw=="}
+        )
 
     def test_none_auth(self, loop):
-        from .py3_only import MockAsyncClient
+
         client = MockAsyncClient(snug.Response(204))
 
-        future = snug.execute_async(myquery(),
-                                    auth=None,
-                                    client=client)
+        future = snug.execute_async(myquery(), auth=None, client=client)
         result = loop.run_until_complete(future)
         assert result == snug.Response(204)
-        assert client.request == snug.GET('my/url')
+        assert client.request == snug.GET("my/url")
 
     def test_auth_callable(self, loop):
-        from .py3_only import MockAsyncClient
-        client = MockAsyncClient(snug.Response(204))
-        auther = methodcaller('with_headers', {'X-My-Auth': 'letmein'})
 
-        future = snug.execute_async(myquery(),
-                                    auth=auther,
-                                    client=client)
+        client = MockAsyncClient(snug.Response(204))
+        auther = methodcaller("with_headers", {"X-My-Auth": "letmein"})
+
+        future = snug.execute_async(myquery(), auth=auther, client=client)
         result = loop.run_until_complete(future)
         assert result == snug.Response(204)
         assert client.request == snug.GET(
-            'my/url', headers={'X-My-Auth': 'letmein'})
+            "my/url", headers={"X-My-Auth": "letmein"}
+        )
 
 
 def test_executor():
-    executor = snug.executor(client='foo')
-    assert executor.keywords == {'client': 'foo'}
+    executor = snug.executor(client="foo")
+    assert executor.keywords == {"client": "foo"}
 
 
 def test_async_executor():
-    executor = snug.async_executor(client='foo')
-    assert executor.keywords == {'client': 'foo'}
+    executor = snug.async_executor(client="foo")
+    assert executor.keywords == {"client": "foo"}
 
 
 def test_relation():
-
     class Foo:
-
         @snug.related
         class Bar(snug.Query):
-            def __iter__(self): pass
+            def __iter__(self):
+                pass
 
             def __init__(self, a, b):
                 self.a, self.b = a, b
 
         class Qux(snug.Query):
-            def __iter__(self): pass
+            def __iter__(self):
+                pass
 
             def __init__(self, a, b):
                 self.a, self.b = a, b
@@ -246,10 +243,7 @@ def test_identity():
     assert snug.query._identity(obj) is obj
 
 
-@py3
 def test__execute_async__(loop):
-    from .py3_only import awaitable
-
     class StringClient:
         def __init__(self, mappings):
             self.mappings = mappings
@@ -259,28 +253,26 @@ def test__execute_async__(loop):
 
     snug.send_async.register(StringClient, StringClient.send)
 
-    client = StringClient({
-        'foo/posts/latest': awaitable('redirect:/posts/latest/'),
-        'foo/posts/latest/': awaitable('redirect:/posts/december/'),
-        'foo/posts/december/': awaitable(b'hello world'),
-    })
+    client = StringClient(
+        {
+            "foo/posts/latest": awaitable("redirect:/posts/latest/"),
+            "foo/posts/latest/": awaitable("redirect:/posts/december/"),
+            "foo/posts/december/": awaitable(b"hello world"),
+        }
+    )
 
     class MyQuery:
-
-        @py2_compatible
         def __iter__(self):
-            redirect = yield '/posts/latest'
-            redirect = yield redirect.split(':')[1]
-            response = yield redirect.split(':')[1]
-            return_(response.decode('ascii'))
+            redirect = yield "/posts/latest"
+            redirect = yield redirect.split(":")[1]
+            response = yield redirect.split(":")[1]
+            return response.decode("ascii")
 
     future = snug.Query.__execute_async__(
-        MyQuery(),
-        client,
-        lambda s: 'foo' + s)
+        MyQuery(), client, lambda s: "foo" + s
+    )
 
-    if sys.version_info > (3, 5):
-        assert inspect.isawaitable(future)
+    assert inspect.isawaitable(future)
 
     result = loop.run_until_complete(future)
-    assert result == 'hello world'
+    assert result == "hello world"