+++ HISTORY.rst
@@ -4,6 +4,14 @@ Release history
 development
 +++++++++++
 
+2.0.0 (2019-10-27)
+++++++++++++++++++
+
+- Add Python 3.8 support
+- Drop Python 2 support
+- Adopt ``black`` autoformatter
+- Fix error on import when no event loop is available
+
 1.4.1 (2019-03-30)
 ++++++++++++++++++